Attempt to fix shader compilation in OSX. strict compilation does not allow dividing float by integer.

+++ source/blender/gpu/shaders/gpu_shader_fx_frag.glsl
@@ -81,7 +81,7 @@ float calculate_ssao_factor(float depth)
         }
     }
 
-    factor /= NUM_SAMPLES * NUM_SAMPLES;    
+    factor /= float(NUM_SAMPLES * NUM_SAMPLES);    
     
     return max(0.0, factor * ssao_params.y);
 }
